--- app/src/main/java/com/example/parkme/Cochera_card_item.kt
@@ -1,8 +1,0 @@
-package com.example.parkme
-
-data class Cochera_card_item (
-        val titulo: String,
-        val direccion: String,
-        val precio: String,
-        val photo: String
-    )
