--- app/src/main/java/com/example/parkme/models/PaginatedResponse.kt
@@ -1,11 +1,0 @@
-package com.example.parkme.models
-
-data class PaginatedResponse<T>(
-    val count: Int,
-    val next: String?,
-    val previous: String?,
-    val results: List<T>
-)
-
-//este es el contrato de la lista ponerse de acuerdo con el team back
-//esto trae la lista de cochera
