+++ app/src/main/java/com/example/parkme/navigation/HistorialFr.kt
@@ -1,19 +1,19 @@
-package com.example.parkme
+package com.example.parkme.navigation
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.example.parkme.R
 
-class PerfilFr : Fragment() {
+class HistorialFr : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_perfil, container, false)
+        return inflater.inflate(R.layout.fragment_historial, container, false)
     }
-
 }
