+++ app/src/main/java/com/example/parkme/navigation/MainActivity.kt
@@ -1,9 +1,10 @@
-package com.example.parkme
+package com.example.parkme.navigation
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.navigation.fragment.NavHostFragment
 import androidx.navigation.ui.NavigationUI
+import com.example.parkme.R
 import com.google.android.material.bottomnavigation.BottomNavigationView
 
 class MainActivity : AppCompatActivity() {
