+++ app/src/main/java/com/example/parkme/navigation/MisCocherasFr.kt
@@ -1,18 +1,19 @@
-package com.example.parkme
+package com.example.parkme.navigation
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.example.parkme.R
 
-class HistorialFr : Fragment() {
+class MisCocherasFr : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_historial, container, false)
+        return inflater.inflate(R.layout.fragment_mis_cocheras, container, false)
     }
 }
