+++ app/src/main/java/com/example/parkme/services/ParkSlotDbService.kt
@@ -1,10 +1,11 @@
 package com.example.parkme.services
 
-import com.example.parkme.database.ParkSlotDbResult
+import com.example.parkme.database.PaginateResponse
+import com.example.parkme.models.ParkSlot
 import retrofit2.Call
 import retrofit2.http.GET
 
 interface ParkSlotDbService {
     @GET("/cocheras")
-    fun listParkSlots(): Call<ParkSlotDbResult>
+    fun listParkSlots(): Call<PaginateResponse<ParkSlot>>
 }
