+++ app/src/main/res/layout/fragment_mis_cocheras.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MisCocherasFr">
+    tools:context=".navigation.MisCocherasFr">
 
     <!-- TODO: Update blank fragment layout -->
